Housekeeping: tidy up layout, css files and components. Remove nav etc for now. Adjust header.
