--- app/components/footer.js
@@ -1,11 +1,0 @@
-import footerStyles from "./footer.module.css";
-
-const Footer = () => (
-  <>
-    <div className={footerStyles.footerContainer}>
-      <p className={footerStyles.footerItem}> © 2023 Jen Corbett</p>
-    </div>
-  </>
-);
-
-export default Footer;
