+++ app/components/layout/footer.module.css
@@ -18,7 +18,6 @@
 @media screen and (min-width: 40rem) {
   .footerContainer {
     justify-content: flex-end;
-
   }
 
   .footerItem {
