+++ app/components/layout/header.js
@@ -1,9 +1,5 @@
-// import Link from "next/link";
-import headerStyles from "@components/header.module.css";
-// import Image from "next/image";
+import headerStyles from "./header.module.css";
 import Head from "next/head";
-
-// const name = "Jen Corbett";
 
 const Header = ({ home, page }) => {
   return (
@@ -11,17 +7,16 @@ const Header = ({ home, page }) => {
       <Head>
         <meta content="" />
       </Head>
-
       <header className={headerStyles.header}>
         {home ? (
           <>
             <div className={`${headerStyles.headerWrapper}`}></div>{" "}
-            <h1 className={headerStyles.heading2Xl}></h1>
+            <h1 className={headerStyles.heading2Xl}>Hey, it's Jen</h1>
           </>
         ) : (
           page && (
             <div>
-              <h1 className={headerStyles.heading2Xl}></h1>
+              <h2 className={headerStyles.headingLg}>Hey, it's Jen</h2>
             </div>
           )
         )}
