+++ app/components/layout/header.module.css
@@ -18,7 +18,19 @@
 }
 
 .heading2Xl {
+  margin-top: 1.5rem;
   font-size: 2rem;
+  line-height: 1.3;
+  font-weight: 800;
+  letter-spacing: -0.05rem;
+  text-align: left;
+  justify-self: center;
+}
+
+
+.headingLg {
+  margin-top: 1.5rem;
+  font-size: 1.5rem;
   line-height: 1.3;
   font-weight: 800;
   letter-spacing: -0.05rem;
