+++ app/components/layout/layout.module.css
@@ -84,8 +84,6 @@ ul.grid {
 
   .grid {
     margin: 0 auto;
-    /* margin-top: 3rem;
-    margin-left: 8rem; */
     align-items: center;
   }
