--- app/components/meta.js
@@ -1,19 +1,0 @@
-import Head from "next/head";
-
-const title = "Jen Corbett Web Development, Design & Digital Strategy";
-
-const Meta = () => (
-  <div>
-    {" "}
-    <Head>
-      {" "}
-      <meta name="description" content="Generated by create next app" />
-      <title>{title}</title>
-      <meta name="viewport" content="width=device-width, initial-scale=1" />
-      <link rel="icon" href="/favicon.ico" />
-      <meta name="og:title" content={title} />{" "}
-    </Head>
-  </div>
-);
-
-export default Meta;
