+++ app/components/utils/navbar.module.css
@@ -14,11 +14,6 @@
 
 }
 
-/* .flex {
-  flex-flow: column wrap;
-  margin: 1rem;
-} */
-
 .icon {
   list-style: none;
   cursor: pointer;
