+++ app/content/posts/tech/hello-world.mdx
@@ -2,7 +2,7 @@
 title: Hello World
 date: 2023-05-10
 summary: A short summary about this post
-image: "https://images.unsplash.com/photo-1620046311691-5d93d65f69e9?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1173&q=80"
+image: ""
 ---
 
 Hi there...
