+++ app/next.config.js
@@ -6,6 +6,9 @@ const nextConfig = {
   // Configure pageExtensions to include md and mdx
   pageExtensions: ["ts", "tsx", "js", "jsx", "md", "mdx", "mjs"],
   // Other Next.js config below
+  experimental: {
+    appDir: true,
+  },
   images: {
     remotePatterns: [
       {
